backend/gen: rename event type
